requirements.txt
Pillow
selenium
pyyaml
